# contracts/openapi/v2/openapi.yaml
openapi: 3.1.0
info:
  title: Resume Matching Platform Java API
  version: 2.1.4
  license:
    name: UNLICENSED
    identifier: LicenseRef-UNLICENSED
  description: |
    Versioned public API owned by the Java service. Resume and match data is
    scoped to the authenticated actor and is never returned in error messages.
servers:
  - url: http://127.0.0.1:8080
    description: Local development server
...
security:
  - bearerAuth: []
paths:
  /api/v2/auth/register:
    post:
      tags: [Authentication]
      summary: Register a user
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v2/auth/login:
    post:
      tags: [Authentication]
      summary: Issue a JWT
      operationId: loginUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/InvalidCredentials'
  /api/v2/auth/password-reset:
    post:
      tags: [Authentication]
      summary: Reset a local account password
      operationId: resetLocalAccountPassword
      security: []
      description: |
        Local-development-only recovery for an existing local account. Java must reject an
        unknown identifier or any non-local account with the same generic `RESOURCE_NOT_FOUND`
        response, without returning account data. This operation never returns a password,
        password hash, token, or account data.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
      responses:
        '204':
          description: Password reset completed. The response has no content.
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
  /api/v2/auth/me:
    get:
      tags: [Authentication]
      summary: Get the current user
      operationId: getCurrentUser
      responses:
        '200':
          description: Current authenticated user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
...
  /api/v2/match-tasks:
    post:
      tags: [Match Tasks]
      summary: Create an evidence-backed match task
      operationId: createMatchTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMatchTaskRequest'
      responses:
        '202':
          description: Match task accepted for asynchronous processing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchTask'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '503':
          $ref: '#/components/responses/ModelUnavailable'
  /api/v2/match-tasks/{taskId}:
    parameters:
      - $ref: '#/components/parameters/TaskId'
    get:
      tags: [Match Tasks]
      summary: Get match task state
      operationId: getMatchTask
      responses:
        '200':
          description: Current asynchronous task state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchTask'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
        '410':
          $ref: '#/components/responses/TaskGone'
  /api/v2/match-tasks/{taskId}/result:
    parameters:
      - $ref: '#/components/parameters/TaskId'
    get:
      tags: [Match Tasks]
      summary: Get a completed match result
      operationId: getMatchTaskResult
      responses:
        '200':
          description: Evidence-backed result for a succeeded task.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
        '409':
          $ref: '#/components/responses/TaskNotReady'
        '410':
          $ref: '#/components/responses/TaskGone'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    TaskId:
      name: taskId
      in: path
      required: true
      description: Match task identifier.
      schema:
        type: string
        pattern: '^task[0-9]{3,}$'
...
  responses:
    BadRequest:
      description: The request is invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthorized:
      description: Authentication is required or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    InvalidCredentials:
      description: The supplied username/email or password did not authenticate. This response is identical for an unknown account and an incorrect password.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InvalidCredentialsError'
...
    ResourceNotFound:
      description: The resource does not exist or is not visible to this actor. Unknown and foreign IDs use this same envelope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResourceNotFoundError'
          example:
            code: RESOURCE_NOT_FOUND
            message: Resource not found.
            correlationId: 00000000-0000-4000-8000-000000000001
            retryable: false
    Conflict:
      description: The request conflicts with existing account state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    IdempotencyConflict:
      description: The idempotency key was reused with different request data.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    UnprocessableEntity:
      description: The request cannot be processed under the current resource state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    ModelUnavailable:
      description: The selected model endpoint is unavailable or timed out.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    TaskGone:
      description: The task was archived, deleted, or otherwise stopped before completion.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    TaskNotReady:
      description: The task has not produced a result.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message, correlationId, retryable]
      properties:
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - AUTHENTICATION_REQUIRED
            - INVALID_CREDENTIALS
            - FORBIDDEN
            - DUPLICATE_RESOURCE
            - INVALID_CONFIRMATION
            - VERSION_CONFLICT
            - RESOURCE_NOT_FOUND
            - RESUME_ARCHIVED
            - RESUME_NOT_EFFECTIVE
            - RESUME_SOFT_DELETED
            - TASK_GONE
            - STALE_ATTEMPT
            - IDEMPOTENCY_CONFLICT
            - MODEL_UNAVAILABLE
            - PYTHON_SERVICE_UNAVAILABLE
            - PYTHON_SERVICE_AUTHENTICATION_FAILED
            - MODEL_OUTPUT_INVALID
            - MODEL_ENDPOINT_REJECTED
            - CALLBACK_DELIVERY_FAILED
            - UNSUPPORTED_FILE
            - PAYLOAD_TOO_LARGE
            - TASK_NOT_READY
          example: RESUME_ARCHIVED
        message:
          type: string
          minLength: 1
          example: The resume is archived and must be restored first.
        correlationId:
          type: string
          format: uuid
        retryable:
          type: boolean
        detailCode:
          type: [string, 'null']
          enum: [DUPLICATE_RESUME_TITLE, null]
        details:
          type: array
          description: Safe validation details without protected content.
          items:
            $ref: '#/components/schemas/ValidationDetail'
      allOf:
        - if:
            required: [detailCode]
            properties:
              detailCode: { const: DUPLICATE_RESUME_TITLE }
          then:
            properties:
              code: { const: DUPLICATE_RESOURCE }
    ValidationDetail:
      type: object
      additionalProperties: false
      required: [field, reason]
      properties:
        field:
          type: string
          minLength: 1
        reason:
          type: string
          minLength: 1
    ResourceNotFoundError:
      allOf:
        - $ref: '#/components/schemas/ApiError'
        - type: object
          required: [code]
          properties:
            code:
              const: RESOURCE_NOT_FOUND
    InvalidCredentialsError:
      allOf:
        - $ref: '#/components/schemas/ApiError'
        - type: object
          required: [code]
          properties:
            code:
              const: INVALID_CREDENTIALS
    UserRole:
      type: string
      enum: [USER, ADMIN]
...
    TaskState:
      type: string
      enum: [QUEUED, PROCESSING, SUCCEEDED, FAILED, TIMED_OUT, BLOCKED]
    SourceType:
      type: string
      enum: [TXT, DOCX]
    JobFamily:
      type: string
      enum: [JAVA_BACKEND]
    RequirementType:
      type: string
      enum: [MANDATORY, PREFERRED]
    MatchStatus:
      type: string
      enum: [SATISFIED, PARTIALLY_SATISFIED, RELATED_BUT_EVIDENCE_INSUFFICIENT, UNMET]
    MatchType:
      type: string
      enum: [EXACT, SEMANTIC, RELATED, NO_MATCH]
    EvidenceStrength:
      type: string
      enum: [NONE, LOW, MEDIUM, HIGH]
    SuggestionState:
      type: string
      enum: [SUPPORTED_FACT, WORDING_ONLY_REWRITE, NEEDS_USER_CONFIRMATION, RISKY_OR_UNSUPPORTED]
    ScoreComponent:
      type: string
      enum: [SKILLS, PROJECT_EXPERIENCE, WORK_CONTENT, EDUCATION_EXPERIENCE, SOFT_SKILLS]
    RegisterRequest:
      type: object
      additionalProperties: false
      required: [username, email, password, role]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 64
          pattern: '^[A-Za-z0-9._-]+$'
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          format: password
          minLength: 12
          maxLength: 128
        role:
          $ref: '#/components/schemas/UserRole'
    LoginRequest:
      type: object
      additionalProperties: false
      required: [identifier, password]
      properties:
        identifier:
          type: string
          minLength: 1
          maxLength: 254
        password:
          type: string
          format: password
          minLength: 1
          maxLength: 128
    PasswordResetRequest:
      type: object
      additionalProperties: false
      required: [identifier, newPassword]
      properties:
        identifier:
          type: string
          minLength: 1
          maxLength: 254
          description: Existing local account username or email. It is not returned by this operation.
        newPassword:
          type: string
          format: password
          minLength: 12
          maxLength: 128
          writeOnly: true
          description: Replacement password. Password confirmation remains a client-only check and is not an API field.
    User:
      type: object
      additionalProperties: false
      required: [id, username, email, role, createdAt]
      properties:
        id:
          type: string
          pattern: '^user[0-9]{3,}$'
        username:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/UserRole'
        createdAt:
          type: string
          format: date-time
    AuthResponse:
      type: object
      additionalProperties: false
      required: [accessToken, tokenType, expiresIn, user]
      properties:
        accessToken:
          type: string
          minLength: 1
        tokenType:
          type: string
          const: Bearer
        expiresIn:
          type: integer
          minimum: 1
          description: Token lifetime in seconds.
        user:
          $ref: '#/components/schemas/User'
...
    CreateMatchTaskRequest:
      type: object
      additionalProperties: false
      required: [resumeId, llmProfileId, jobFamily, jobDescriptionText, idempotencyKey]
      properties:
        resumeId:
          type: string
          pattern: '^resume[0-9]{3,}$'
        llmProfileId:
          type: string
          pattern: '^profile[0-9]{3,}$'
        jobFamily:
          $ref: '#/components/schemas/JobFamily'
        jobDescriptionText:
          type: string
          minLength: 20
          maxLength: 20000
        idempotencyKey:
          type: string
          minLength: 16
          maxLength: 128
    MatchTask:
      type: object
      additionalProperties: false
      required: [id, resumeId, llmProfileId, jobFamily, state, attempt, resumeVersion, createdAt, updatedAt]
      properties:
        id:
          type: string
          pattern: '^task[0-9]{3,}$'
        resumeId:
          type: string
          pattern: '^resume[0-9]{3,}$'
        llmProfileId:
          type: string
          pattern: '^profile[0-9]{3,}$'
        jobFamily:
          $ref: '#/components/schemas/JobFamily'
        state:
          $ref: '#/components/schemas/TaskState'
        attempt:
          type: integer
          minimum: 1
        resumeVersion:
          type: integer
          minimum: 0
        failureCode:
          type: [string, 'null']
          enum: [MODEL_UNAVAILABLE, PYTHON_SERVICE_UNAVAILABLE, PYTHON_SERVICE_AUTHENTICATION_FAILED, MODEL_OUTPUT_INVALID, MODEL_ENDPOINT_REJECTED, CALLBACK_DELIVERY_FAILED, UNSUPPORTED_FILE, TASK_GONE, STALE_ATTEMPT, null]
        resultAvailable:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Evidence:
      type: object
      additionalProperties: false
      required: [id, sourceType, sourceLocation, excerpt, confidence, strength]
      properties:
        id:
          type: string
          pattern: '^evidence[0-9]{3,}$'
        sourceType:
          $ref: '#/components/schemas/SourceType'
        sourceLocation:
          type: string
          minLength: 1
          description: Stable section, paragraph, or experience location in the resume.
        sourceStart:
          type: integer
          minimum: 0
        sourceEnd:
          type: integer
          minimum: 0
        excerpt:
          type: string
          minLength: 1
        confidence:
          type: number
          minimum: 0
          maximum: 1
        strength:
          $ref: '#/components/schemas/EvidenceStrength'
    RequirementMatch:
      type: object
      additionalProperties: false
      required: [requirementId, requirementText, requirementType, matchStatus, matchType, component, componentScore, evidence, gap, suggestionState]
      properties:
        requirementId:
          type: string
          pattern: '^requirement[0-9]{3,}$'
        requirementText:
          type: string
          minLength: 1
        requirementType:
          $ref: '#/components/schemas/RequirementType'
        matchStatus:
          $ref: '#/components/schemas/MatchStatus'
        matchType:
          $ref: '#/components/schemas/MatchType'
        component:
          $ref: '#/components/schemas/ScoreComponent'
        componentScore:
          type: number
          minimum: 0
          maximum: 1
        evidence:
          type: array
          items:
            $ref: '#/components/schemas/Evidence'
        gap:
          type: [string, 'null']
        suggestionState:
          $ref: '#/components/schemas/SuggestionState'
      allOf:
        - if:
            properties:
              matchStatus:
                enum: [SATISFIED, PARTIALLY_SATISFIED]
          then:
            properties:
              evidence:
                minItems: 1
        - if:
            properties:
              suggestionState:
                enum: [SUPPORTED_FACT, WORDING_ONLY_REWRITE]
          then:
            properties:
              evidence:
                minItems: 1
    MatchSuggestion:
      type: object
      additionalProperties: false
      required: [id, requirementId, state, proposedText, evidenceIds]
      properties:
        id:
          type: string
          pattern: '^suggestion[0-9]{3,}$'
        requirementId:
          type: string
          pattern: '^requirement[0-9]{3,}$'
        state:
          $ref: '#/components/schemas/SuggestionState'
        originalText:
          type: [string, 'null']
        proposedText:
          type: string
          minLength: 1
        evidenceIds:
          type: array
          items:
            type: string
            pattern: '^evidence[0-9]{3,}$'
        requiresUserConfirmation:
          type: boolean
          default: false
      allOf:
        - if:
            properties:
              state:
                enum: [SUPPORTED_FACT, WORDING_ONLY_REWRITE]
          then:
            properties:
              evidenceIds:
                minItems: 1
    ScoreBreakdown:
      type: object
      additionalProperties: false
      required: [skills, projectExperience, workContent, educationExperience, softSkills, composite]
      properties:
        skills:
          type: number
          minimum: 0
          maximum: 1
          description: Fixed weight 0.40.
        projectExperience:
          type: number
          minimum: 0
          maximum: 1
          description: Fixed weight 0.25.
        workContent:
          type: number
          minimum: 0
          maximum: 1
          description: Fixed weight 0.15.
        educationExperience:
          type: number
          minimum: 0
          maximum: 1
          description: Fixed weight 0.10.
        softSkills:
          type: number
          minimum: 0
          maximum: 1
          description: Fixed weight 0.10.
        composite:
          type: number
          minimum: 0
          maximum: 1
    MatchResult:
      type: object
      additionalProperties: false
      required: [taskId, resumeId, resumeVersion, jobDescriptionText, score, requirements, suggestions, completedAt]
      properties:
        taskId:
          type: string
          pattern: '^task[0-9]{3,}$'
        resumeId:
          type: string
          pattern: '^resume[0-9]{3,}$'
        resumeVersion:
          type: integer
          minimum: 0
        jobDescriptionText:
          type: string
          minLength: 20
          maxLength: 20000
        score:
          $ref: '#/components/schemas/ScoreBreakdown'
        requirements:
          type: array
          items:
            $ref: '#/components/schemas/RequirementMatch'
        suggestions:
          type: array
          items:
            $ref: '#/components/schemas/MatchSuggestion'
        completedAt:
          type: string
          format: date-time